Moved VectorInput into the lib. Refactored a bit
