+++ navlie/__init__.py
@@ -4,7 +4,6 @@ from .types import (
     MeasurementModel,
     ProcessModel,
     Input,
-    VectorInput,
     StateWithCovariance,
     Dataset,
 )
@@ -40,3 +39,5 @@ from .utils import (
     randvec,
     jacobian,
 )
+
+from .lib import VectorInput  # for backwards compatibility
