+++ tests/unit/test_body_velocity.py
@@ -1,13 +1,12 @@
-from navlie.lib.states import SE2State, SE3State, SE23State
-from navlie.lib.models import (
+from navlie.lib import SE2State, SE3State
+from navlie.lib import (
     BodyFrameVelocity,
-    RangePoseToAnchor,
     RelativeBodyFrameVelocity,
+    VectorInput,
 )
-from pymlg import SO2, SO3, SE3, SE2, SE3, SE23
+from pymlg import SE3, SE2, SE3
 import numpy as np
 import pytest
-from navlie.types import VectorInput
 
 
 @pytest.mark.parametrize("direction", ["left", "right"])
